Make nullonerror the default
nullonerror controls that the jsonptr_get_*() functions will
suppress an error if the jsonvalue requested cannot be casted
to the requested data type (like '123.4' cannot be converted
to int8). nullonerror=true will return NULL instead of causing
a PostgreSQL ERROR exception. This is the desired default
behavior when using these functions inside the expressions for
indexes.